[refactor] 클리어/실패 모달 로직 분리 및 연결
